requirements.txt
backports.shutil-get-terminal-size==1.0.0
cycler==0.10.0
decorator==4.0.9
Django==1.9.7
django-annoying==0.9.0
django-cors-headers==1.1.0
docutils==0.12
httplib2==0.9.2
ipython==4.2.0
ipython-genutils==0.1.0
lxml==3.6.0
matplotlib==1.5.1
numpy==1.11.0
pandas==0.18.1
pathlib2==2.1.0
pdfkit==0.5.0
pickleshare==0.7.2
Pillow==3.2.0
progressbar==2.3
psycopg2==2.6.2
Pygments==2.1.3
pyparsing==2.1.1
pypiwin32==219
python-dateutil==2.5.3
pytz==2016.4
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
Twisted==16.4.0
virtualenv==15.0.2
virtualenvwrapper-win==1.2.1
xlrd==1.0.0
zope.interface==4.2.0
